refactor(frontend): move plan chip to user dropdown
Move the subscription tier chip from the top bar into
the user dropdown menu, displayed under the username.
The "Get Premium" button remains in the top bar for
non-premium users.
