+++ frontend/app/src/components/premium/GetPremiumButton.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import { externalLinks } from '@shared/external-links';
 import ExternalLink from '@/components/helper/ExternalLink.vue';
 import { usePremiumHelper } from '@/composables/premium';
 
@@ -8,17 +7,15 @@ const { hideOnSmallScreen = false } = defineProps<{
 }>();
 
 const { t } = useI18n({ useScope: 'global' });
-const { currentTier, premium } = usePremiumHelper();
+const { premium } = usePremiumHelper();
 const { isLgAndDown } = useBreakpoint();
-
-const tierLabel = computed<string>(() => {
-  const tier = get(currentTier);
-  return tier ? t('premium_placeholder.current_plan', { tier }) : t('premium_placeholder.upgrade_plan');
-});
 </script>
 
 <template>
-  <div class="mr-2">
+  <div
+    v-if="!premium"
+    class="mr-2"
+  >
     <RuiTooltip
       :popper="{ placement: 'bottom' }"
       :disabled="!(isLgAndDown && hideOnSmallScreen)"
@@ -26,7 +23,6 @@ const tierLabel = computed<string>(() => {
     >
       <template #activator>
         <ExternalLink
-          v-if="!premium"
           custom
           premium
         >
@@ -43,23 +39,8 @@ const tierLabel = computed<string>(() => {
             {{ t('premium_settings.get') }}
           </RuiButton>
         </ExternalLink>
-        <ExternalLink
-          v-else
-          custom
-          :url="externalLinks.manageSubscriptions"
-        >
-          <RuiChip
-            size="sm"
-            variant="outlined"
-            color="primary"
-            data-cy="get-premium-button"
-            class="!cursor-pointer"
-          >
-            {{ tierLabel }}
-          </RuiChip>
-        </ExternalLink>
       </template>
-      <span>{{ premium ? tierLabel : t('premium_settings.get') }}</span>
+      <span>{{ t('premium_settings.get') }}</span>
     </RuiTooltip>
   </div>
 </template>
